feat!: make prepare ESM-first
BREAKING CHANGE: bunchee prepare now generates ESM-only package.json
configuration by default. Use prepare --cjs to opt in to dual CJS+ESM
output. The --esm flag is replaced by --cjs (inverted default).

- type: module is set by default (previously only with --esm)
- exports use types/default conditions without import/require split
- build scripts are always added (previously only with --esm)
