+++ test/integration/prepare-js/package.json
