+++ test/integration/prepare-ts-cjs/.gitignore
@@ -1,0 +1,2 @@
+package.json
+tsconfig.json
