+++ test/integration/prepare-ts-cjs/src/bin/cli.ts
