+++ test/integration/prepare-ts-cjs/src/bin/cmd.ts
