+++ test/integration/prepare-ts-cjs/src/foo.ts
