+++ test/integration/prepare-ts-cjs/src/index.react-server.ts
