+++ test/integration/prepare-ts-cjs/src/index.ts
