--- test/integration/prepare-ts-esm/.gitignore
@@ -1,5 +1,0 @@
-dist/
-tsconfig.tsbuildinfo
-package.json
-tsconfig.json
-
