+++ test/integration/prepare-ts/.gitignore
@@ -1,2 +1,5 @@
+dist/
+tsconfig.tsbuildinfo
 package.json
 tsconfig.json
+
