+++ test/integration/prepare-ts/src/bin.ts
